Feature: Integrated Abstract
Made some classes abstract.
