+++ src/structures/commands-map.ts
@@ -1,10 +1,11 @@
-import { ApplicationCommandType } from "./application-commands";
+import { MessageCommand, UserCommand } from "./menu-command";
+import { Command } from "./slash-command";
 
 /**
  * Commands Map.
  * Extends Map to easily access commands; store and get commands.
  */
-export default class CommandsMap extends Map<string, ApplicationCommandType>{
+export default class CommandsMap extends Map<string, Command | UserCommand | MessageCommand>{
   /**
    * Get an array of raw data of all commands.
    */
