--- src/structures/field-builder.ts
@@ -1,16 +1,0 @@
-import Field from "./field";
-import { choice } from "./field";
-
-/**
- * Managed field structure for creating field types.
- */
-export default class FieldBuilder extends Field {
-  constructor(
-    readonly name: string,
-    readonly description: string,
-    readonly required: boolean = false,
-    readonly choices: choice[] = []
-  ) {
-    super();
-  }
-}
