--- src/structures/index.ts
@@ -1,5 +1,0 @@
-export * from "./application-commands";
-export * from "./slash-command-consumer";
-export * from "./menu-command-consumer";
-export * from "./field-consumer";
-export * from "./interactions-handler";
