+++ src/structures/menu-command.ts
@@ -1,22 +1,28 @@
+import { ContextMenuInteraction } from "discord.js";
+
 /**
  * Interface for context menu command creation.
  */
 export default interface ContextMenuCommand {
   /**
-   * The type of context menu command.
-   */
-  type: number;
-  /**
    * The name of the command
    */
   name: string
-};
+  /**
+   * The function to execute when command is called by context menu
+   * @param interaction The context menu interaction object
+   */
+  execute(interaction: ContextMenuInteraction): void
+}
 
 /**
- * Base structure context menu command types.
- * Do not use for creating context menu commands.
+ * Managed class for creating context menu command types.
  */
-export default class ContextMenuCommand {
+export default abstract class ContextMenuCommandBuilder {
+  /**
+   * The type of context menu command.
+   */
+  abstract type: number;
   /**
    * The raw object for context menu command data. Used to interact with discord.
    */
@@ -26,4 +32,14 @@ export default class ContextMenuCommand {
       type: this.type
     })
   }
+}
+
+/** Context menu user command */
+export abstract class UserCommand extends ContextMenuCommandBuilder {
+  readonly type = 2;
+}
+
+/** Context menu message command */
+export abstract class MessageCommand extends ContextMenuCommandBuilder {
+  readonly type = 3;
 }
