+++ src/structures/slash-command.ts
@@ -1,11 +1,10 @@
+import { CommandInteraction } from "discord.js";
+import { FieldType } from "./field";
+
 /** 
  * Interface for slash command creation.
  */
-export default interface SlashCommand {
-  /**
-   * Type of command or option.
-   */
-  type: number;
+export default interface SlashCommandBuilder<TOptions extends any[]> {
   /**
    * The name of the command.
    */
@@ -13,18 +12,33 @@ export default interface SlashCommand {
   /**
    * The description of the command.
    */
-  description: string,
+  description: string
   /**
-   * The options of the command.
+   * The function to execute when command is called
+   * @param interaction The command interaction object
    */
-  options?: any[];
+  execute(interaction: CommandInteraction): void;
 }
 
 /**
- * Base structure slash command structure.
- * Do not use for creating commands.
+ * Managed slash command structure for creating slash command types(have options property).
  */
-export default class SlashCommand {
+export default abstract class SlashCommandBuilder<TOptions extends any[]> {
+  /**
+   * Type of command or option.
+   */
+  abstract type: number;
+  /**
+   * The mapped options.
+   * Used for easily finding options using option name.
+   */
+  readonly options: Map<string, TOptions[number]>
+  /**
+   * @param _options The options for the command.
+   */
+  constructor(readonly _options: TOptions | [] = []) {
+    this.options = new Map(_options.map(option => [option.name, option]));
+  }
   /**
    * The raw object for command data. Used to interact with discord.
    */
@@ -32,8 +46,23 @@ export default class SlashCommand {
     return ({
       name: this.name,
       description: this.description,
-      options: this.options,
+      options: this._options.map(option => option.rawData),
       type: this.type
     })
   }
+}
+
+/** Subcommand */
+export abstract class Subcommand extends SlashCommandBuilder<FieldType[]> {
+  readonly type = 1;
+}
+
+/** Subcommand Group */
+export abstract class SubcommandGroup extends SlashCommandBuilder<Subcommand[]> {
+  readonly type = 2;
+}
+
+/** Command */
+export abstract class Command extends SlashCommandBuilder<(SubcommandGroup | Subcommand)[] | FieldType[]> {
+  readonly type = 1;
 }
